The AI works now!
One problem my AI encountered earlier was that it was trying to bust
ghosts which were already being contained or had been released.

To recitfy this, I removed a ghost from the list of ghosts whenever I
found that, while updating the fields for a buster, the state of the
buster had changed (indicating that it had just picked up a ghost)

In order to do this, I used an Iterator loop copy pasted from stack
overflow, which works nicely. I did this, because simply using a foreach
loop probably wouldn't allow me to remove elements from the list while
iterating. This is the first time I've ever used Iterator, so this was a
good opportunity.

New problem: my AI needs a better ghostfinding algorithm.
---------------------------------------------------------

Currently, it has no problem clearing the middle of the arena, since the
busters are programmed to go there when they can't see any ghosts, but it
has trouble reaching any ghosts outside of that small zone in the middle.

Two solutions which immediately occur are:
	1. Hardcode a 'patrolling path' which allows the buster to cover
	the entire arena with its view range
	2. Randomly generate destination coordinates and pray to RNGesus

Another more complicated solution (suggested by your fellow competitors)
involves the use of a matrix to keep track of the probability that a ghost
is located in a certain region of the arena, based on previous sightings
of the ghost:

	GideonHoogeveen 00:13
	I'm building a matrix that breaks the game up in tiles and
	calculates the suspicion that a ghost is there for each tile.

	GideonHoogeveen 00:14
	when you visit a tile there is no ghost so -100, when you see a
	ghost it will be 100

	and it normalizes over time