About to hardcode a patrolling path
After doing some maths, I figured the approach I'd take is to make a
coordinate class (or vector class whatever) and a matrix class to store a
list of coordinates (so I can use matrix coordinates to refer to map
coordinates). Actually, maybe the matrix doesn't have to be a class,
though perhaps a class is a better idea, because I can chuck in some
methods to convert between matrix and map coordinates

I figured I'd need to store 9 x 18 coordinates in this matrix: this is
based off some rough calculations assuming that the busters can only move
up, down, left, or right by 800 at a time, and how far they need to move
to at least cover the entire map with their view range. There may be a
more efficient way to cover the entire map using some diagonal movement
(especially in the corners), though that is probably too complicated.

Once that is done, I will store a list of coordinates that will form the
patrol path. Actually, I'll have two lists to store two patrol paths, one
path will be a spiral covering the top half of the arena, another spiral
path will cover the bottom. Which path a buster will take depends on
whether it's iteration index is odd or even.

I may later experiment with giving the busters a 'destination/target'
field, which might work if you check if the buster has reached target yet
(if yes, change target, if not, keep going, something like that)

I suppose I could change the x y fields in my entities to Coordinates, but
ceebs