Fixed some strange behaviours, but there's more
Look at this:
> BUST 19
> BUST 19
> BUST 19
> BUST 19

Buster 0 of player blueCicada trapped ghost 19.
Buster 1 of player blueCicada wants to bust ghost 19 but can't see it.
Buster 2 of player blueCicada wants to bust ghost 19 but can't see it.
Buster 3 of player blueCicada wants to bust ghost 19 but can't see it.

One way to fix this behaviour is to access the value of the target ghost
to see how many busters are trying to trap the ghost, and hence calculate
when the others should stop, though this might be a little complicated
It's not just as simple as popping the ghost off the list after busting
it, now that it takes multiple busts to capture a ghost. You have to
figure out when and who stops busting, and who gets to keep the ghost.

On second thought, there may be a slightly simpler way of doing this: if
each buster keeps track of who else is busting the same ghost (this can be
done by looking at both the current iteration index and the ghost value to
deduce which busters before it may possibly have captured the ghost), we
could look at the value of the busters before it (who were also busting
that ghost) to check if the ghost has been captured. This is feasible, but
might take some effort so I'll put it off until a little later.