Found and (sort of) fixed dead-end in logic
Inside the 'chase or bust ghost' control block, there is no else statement
to catch the case where the ghost is too close, ie. not within busting
range, but within sighting range.

I thought that iterating through the list to search for other potential
targets would work, but instead, the AI just started acting really stupid,
so I was forced to undo that. One possible explanation for that was that
because I used a foreach loop to iterate through an iterable, the order of
the objects was not preserved during iteration, causing the buster to run
all over the place. I didn't actually check to see if the busters were
actually running between ghosts, just guessing here.

My quick, albeit dodgy fix was just to default to MOVE 8500 4500 if no
other decision could be made, and that earned (probably) my first victory
against Bosslime.

You may want to consider trying the for loop idea again later, maybe using
the regular for loop. Or, just take the next ghost, if there is one.

EDIT: As of submitting this code, I have been promoted to the Silver
League (woot!)

At this point, I should make a list of things I need to improve:

- A better default action than MOVE 8500 4500
- Logic for endgame play: eg. chase and stun enemy buster carrying a
  ghost to their base if they are not close enough to release the ghost
  Or just chase a visible foe if you don't know who is carrying the last
  ghosts
  Or go to the last known location of a buster or head near their base if
  you can't see any foes
- A better way of choosing a ghost or buster to target
  	For targeting busters:
	Consider whether they're carrying a ghost (while not close enough
	to release the ghost to their base)
	Also consider how far away they are from you
	You may even want to consider stunning the one with the least
	cooldown time remaining
	For targeting ghosts:
	Consider their stamina (a ghost with a fuckton of stamina is worth
	just as much as one with very little stamina)
	And their distance from you (and possibly distance from the base)
- A fix for busters trying to bust ghosts that were trapped by a previous
  buster in the same turn. If we're going with the idea of keeping
  track of the busters busting the same ghost, we could potentially make a
  new field for the ghost, a list which stores the IDs of the busters
  busting it. Each time a buster wants to bust a ghost, it checks to see
  if any of the busters with IDs in that list are now carrying a ghost (do
  we need a mechanism for checking if it didn't just randomly bust another
  ghost with 0 stamina on that turn??? Maybe a field which only ever
  contains a ghostID if it was busted in the current turn?).
  To make this a little more efficient, consider checking only the busters
  that moved before you
  (this means that the first buster does not need to do any of this
  checking, except maybe in the case there are also enemies trying to bust
  that ghost)
  Actually, I just figured out another way: give the ghost a 'busted'
  flag, so other busters can use it to check if it was busted in the
  current turn. The only time this won't work is if the opponent manages
  to bust the ghost before you do. Perhaps we could implement a hybrid of
  these methods?
- Parasite strategy: Steal ghosts that have been weakened by a foe
  You may even want to ignore ghosts with stamina beyond a certain
  threshold and let your enemies work at it (this threshold can likely
  only be refined through trial and error)
- Don't have multiple busters stun a single foe: this is extremely
  wasteful, since is it proven that the waiting times do not stack.
  (however, it is uncertain whether successive stuns reset the counter
  back to ten turns)
  One way you might do this is have a field which serves as a flag: it
  only contains a foe's buster ID if it was stunned in the current turn.
  Then, whenever an allied buster wants to stun a foe, it looks for that
  foe's ID in the stunFlags of it's allies.

^These suggestions were written with the assumption that stunning a buster
who is carrying a ghost within release range of their base would cause
them to claim the point for the ghost, though this may not be what
actually happens - there is conflicting information being shared in the
public chat

You may also want to reconsider whether allowing busters to stun while
carrying is a good idea - while you can pick up the dropped ghost again,
you need some distance before you can do that, and this may not be
practical.

We are currently in the mid-lower end of the Silver League: this would be
a good point to revert to if you fuck up anything with experimental
changes.