Minor change
Just implemented the busted flag: this will hopefully solve part of the
problem of trying to bust a ghost that was just captured in the current
turn by a previous buster. However, complications will likely arise if
enemies are also competing for the ghost.

As mentioned before, you may want to resolve this by looking at the number
of busters busting that ghost. If that number exceeds the number of allied
busters busting that ghost, that means at least one enemy is also trying
to trap the ghost.

After rereading the rules, it seems that the team with most busters will
take the ghost, and the closest buster from that team will take the ghost.
However, I think my system of setting the flag works just as well, don't
need to calculate distance. However, you may need to calculate distance if
you suspect the enemy may have also taken the ghost - this will only
happen when the enemy team has more busters working on the ghost than you
do