Fix doesn't work
After a bit of debugging, it seems that the fix can't work. It's not a
matter of who busts the ghost first, it's about who is the closest to the
ghost (see the rules). For this reason, it pays to have an array field in
the Ghost which keeps track of who is currently busting it, so you can
then iterate through to see who is the closest.

I will likely remove the busted field in an upcoming commit