Setting threshold to 0
I imagined doing this would've made it functionally the same as what I had
before I reworked the ghost storage, but from running this code against my
older code, it seems that this seems to work better somehow, so I'll
submit it now and see how it goes.

EDIT: On second inspection, no - it's doing worse. The current best
performing version of your code is under the commit titled 'Adding a
tentative solution to the multiple stun problem'