GOLD LEAGUE WOT
I made a tiny adjustment to the starting points of my patrol sprials and
I'm now in gold, wot

Also in this commit, but commented out because I want to save it for the
next commit is a fix in progress for allies trying to stun outdated
locations. For this, I'm bringing back the radarcount field, but this time
for foes, not ghosts.