Implemented PQ for ghosts
Using stamina comparator
Hope to god this works
I'm not sure if I'm sorting by ascending or descending stamina, but it
should be ascending towards the end of the list

EDIT: Testing to see whether the stun while carrying ghost thing was
effective was inconclusive, so I'll leave it as "don't stun while
carrying"

Meanwhile, I wait for the version containing the PQ to be tested

Another thing you could possibly do is make a comparator that takes into
account both the distance and the stamina of a ghost. One way you might do
this is to directly compare the products of the ghost's distance and
stamina, or, consider the fact that a turn is worth up to one stamina
point or 800 steps, so convert stamina into units of distance by *800.

So we can try two different comparators:
ProductComparator, and CostComparator