Discovered that I'm not handing control from romka to dumbai properly
If the control flow reaches the block where romka hands control over to
dumb ai, a frameshift in the output is guaranteed to occur
To remedy this, I might reorder the control flow, or simply merge the
romka method into dumbai

EDIT: Brain dumping some high level ideas
You may want to investigate the possibility of hearding ghosts back after
returning from the rush. You will move slower, but if there are enough
ghosts, you could profit immensely.
One possible way of counting ghosts that could potential be herded is by
making a list of ghost ids you encounter, and by the end of the rush, ids
belonging to ghosts which lie outside the rush radius are culled from the
list, leaving you with an idea of how many ghosts you might be able to
scrape in.
There is, however, a decent chance you will collide with the enemy in the
centre, and if you trade stuns, this will break the formation necessary
for herding, especially if one of the central busters is affected, so
maybe check for that.